Fixing compiler issues on Windows.

+++ libNeonDomain/include/Neon/domain/details/dGrid/dGrid_imp.h
@@ -228,7 +228,7 @@ auto dGrid::newContainer(const std::string& name,
                                                                      *this,
                                                                      lambda,
                                                                      defaultBlockSize,
-                                                                     [](const Neon::index_3d&) { return size_t(0); });
+                                                                     [](const Neon::index_3d&) { return 0; });
     return c;
 }
