+++ libNeonDomain/include/Neon/domain/tools/gridTransformer/tGrid_ti.h
@@ -129,7 +129,7 @@ auto tGrid<GridTransformation>::newContainer(const std::string& name,
                                                                                *this,
                                                                                lambda,
                                                                                defaultBlockSize,
-                                                                               [](const Neon::index_3d&) { return size_t(0); });
+                                                                               [](const Neon::index_3d&) { return 0; });
     return kContainer;
 }
